Move API server files into API folder
